feat: migrate data layer to modular architecture
- Created core:data, core:database, and core:domain modules
- Moved repositories, DAOs, entities, and use cases from app module
- Updated imports and module dependencies across project
- Registered new modules in settings.gradle.kts
