--- app/src/main/java/com/example/petbreeds/di/RepositoryModule.kt
@@ -1,20 +1,0 @@
-package com.example.petbreeds.di
-
-import com.example.petbreeds.data.repository.PetRepositoryImpl
-import com.example.petbreeds.domain.repository.PetRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class RepositoryModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindPetRepository(
-        petRepositoryImpl: PetRepositoryImpl
-    ): PetRepository
-}
