+++ app/src/main/java/com/example/petbreeds/domain/usecase/GetPetsUseCase.kt
@@ -3,7 +3,7 @@ package com.example.petbreeds.domain.usecase
 import com.example.common.NetworkResult
 import com.example.model.Pet
 import com.example.model.PetType
-import com.example.petbreeds.domain.repository.PetRepository
+import com.example.domain.repository.PetRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
