+++ app/src/main/java/com/example/petbreeds/domain/usecase/ToggleFavoriteUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.petbreeds.domain.usecase
 
-import com.example.petbreeds.domain.repository.PetRepository
+import com.example.domain.repository.PetRepository
 import javax.inject.Inject
 
 class ToggleFavoriteUseCase @Inject constructor(
