+++ app/src/main/java/com/example/petbreeds/presentation/MainActivity.kt
@@ -1,6 +1,7 @@
 package com.example.petbreeds.presentation
 
 import android.os.Bundle
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.padding
@@ -13,6 +14,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.example.petbreeds.core.ui.theme.PetBreedsTheme
 import com.example.model.PetType
+import com.example.petbreeds.core.network.BuildConfig
 import com.example.petbreeds.presentation.navigation.BottomNavItem
 import com.example.petbreeds.presentation.navigation.PetBreedsNavigation
 import com.example.petbreeds.presentation.navigation.Routes
@@ -29,6 +31,7 @@ class MainActivity : ComponentActivity() {
     @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         setContent {
             val navController = rememberNavController()
             val currentPetType by preferencesManager.petTypeFlow.collectAsState(initial = null)
