+++ app/src/test/java/com/example/petbreeds/data/mapper/PetMappersTest.kt
@@ -3,7 +3,7 @@ package com.example.petbreeds.data.mapper
 import com.example.petbreeds.data.api.dto.CatBreedDto
 import com.example.network.dto.DogBreedDto
 import com.example.petbreeds.data.api.dto.ImageDto
-import com.example.petbreeds.data.local.entity.PetEntity
+import com.example.database.entity.PetEntity
 import com.example.model.PetType
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
