+++ core/data/src/main/java/com/example/data/repository/PetRepositoryImpl.kt
@@ -1,0 +1,110 @@
+package com.example.data.repository
+
+import com.example.common.NetworkResult
+import com.example.data.mapper.toCatEntities
+import com.example.data.mapper.toDogEntities
+import com.example.data.mapper.toDomain
+import com.example.database.dao.PetDao
+import com.example.domain.repository.PetRepository
+import com.example.model.Pet
+import com.example.model.PetType
+import com.example.network.service.CatApiService
+import com.example.network.service.DogApiService
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.map
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class PetRepositoryImpl @Inject constructor(
+    private val catApiService: CatApiService,
+    private val dogApiService: DogApiService,
+    private val petDao: PetDao
+) : PetRepository {
+
+    override fun getPets(petType: PetType): Flow<NetworkResult<List<Pet>>> {
+        return petDao.getPetsByType(petType).map { entities ->
+            NetworkResult.Success(entities.toDomain())
+        }
+    }
+
+    override fun getFavoritePets(petType: PetType): Flow<List<Pet>> {
+        return petDao.getFavoritePetsByType(petType).map { entities ->
+            entities.toDomain()
+        }
+    }
+
+    override suspend fun refreshPets(petType: PetType, page: Int, query: String?): NetworkResult<Unit> {
+        return try {
+            // Fetch data from API
+            val pets = when (petType) {
+                PetType.CAT -> {
+                    val response = if (!query.isNullOrBlank()) {
+                        catApiService.searchBreeds(query)
+                    } else {
+                        catApiService.getBreeds(limit = 20, page = page)
+                    }
+                    response.toCatEntities()
+                }
+                PetType.DOG -> {
+                    val response = if (!query.isNullOrBlank()) {
+                        dogApiService.searchBreeds(query)
+                    } else {
+                        dogApiService.getBreeds(limit = 20, page = page)
+                    }
+                    response.toDogEntities()
+                }
+            }
+
+            // Handle different scenarios
+            when {
+                // Search query - replace all data
+                !query.isNullOrBlank() -> {
+                    val favoriteIds = petDao.getFavoritePetsByType(petType).first().map { it.id }.toSet()
+                    val petsWithFavorites = pets.map { newPet ->
+                        newPet.copy(isFavorite = favoriteIds.contains(newPet.id))
+                    }
+                    petDao.refreshPetsForFirstPage(petsWithFavorites, petType)
+                }
+
+                // First page - replace all data
+                page == 0 -> {
+                    val favoriteIds = petDao.getFavoritePetsByType(petType).first().map { it.id }.toSet()
+                    val petsWithFavorites = pets.map { newPet ->
+                        newPet.copy(isFavorite = favoriteIds.contains(newPet.id))
+                    }
+                    petDao.refreshPetsForFirstPage(petsWithFavorites, petType)
+                }
+
+                // Subsequent pages - append only new data
+                else -> {
+                    val existingPets = petDao.getPetsByType(petType).first()
+                    val favoriteIds = existingPets.filter { it.isFavorite }.map { it.id }.toSet()
+
+                    val petsWithFavorites = pets.map { newPet ->
+                        newPet.copy(isFavorite = favoriteIds.contains(newPet.id))
+                    }
+
+                    petDao.appendPets(petsWithFavorites, petType)
+                }
+            }
+
+            NetworkResult.Success(Unit)
+        } catch (e: Exception) {
+            NetworkResult.Error(e.message ?: "Unknown error occurred")
+        }
+    }
+
+    override suspend fun toggleFavorite(petId: String) {
+        val pet = petDao.getPetById(petId)
+        pet?.let {
+            val newFavoriteStatus = !it.isFavorite
+            petDao.updateFavoriteStatus(petId, newFavoriteStatus)
+        }
+    }
+
+    override suspend fun getPetDetails(petId: String): Pet? {
+        return petDao.getPetById(petId)?.toDomain()
+    }
+}
