+++ core/database/src/main/java/com/example/database/PetDatabase.kt
@@ -1,0 +1,17 @@
+package com.example.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import androidx.room.TypeConverters
+import com.example.database.dao.PetDao
+import com.example.database.entity.PetEntity
+
+@Database(
+    entities = [PetEntity::class],
+    version = 2,
+    exportSchema = false
+)
+@TypeConverters(Converters::class)
+abstract class PetDatabase : RoomDatabase() {
+    abstract fun petDao(): PetDao
+}
