+++ gradle/libs.versions.toml
@@ -1,8 +1,10 @@
 [versions]
 # Build
 agp = "8.13.0"
+hilt-android = "2.57.2"
 kotlin = "1.9.22"
 ksp = "1.9.22-1.0.17"
+
 
 # AndroidX
 core-ktx = "1.12.0"
@@ -37,9 +39,11 @@ truth = "1.1.3"
 turbine = "0.12.1"
 appcompat = "1.6.1"
 material = "1.10.0"
+databinding-adapters = "8.13.0"
 
 [libraries]
 # Kotlin
+hilt-android-v2572 = { module = "com.google.dagger:hilt-android", version.ref = "hilt-android" }
 kotlin-stdlib = { module = "org.jetbrains.kotlin:kotlin-stdlib", version.ref = "kotlin" }
 coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-android", version.ref = "coroutines" }
 coroutines-test = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-test", version.ref = "coroutines" }
@@ -94,6 +98,7 @@ android-gradle-plugin = { module = "com.android.tools.build:gradle", version.ref
 kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin" }
 androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
 material = { group = "com.google.android.material", name = "material", version.ref = "material" }
+androidx-databinding-adapters = { group = "androidx.databinding", name = "databinding-adapters", version.ref = "databinding-adapters" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
