+++ settings.gradle.kts
@@ -24,3 +24,6 @@ include(":app")
 include(":core:model")
 include(":core:common")
 include(":core:network")
+include(":core:database")
+include(":core:data")
+include(":core:domain")
